backend: Split thirty year regulation tests into multiple files
